+++ circus-train-s3-mapreduce-cp/src/main/java/com/hotels/bdp/circustrain/s3mapreducecp/ConfigurationVariable.java
@@ -32,6 +32,7 @@ final class Constants {
 
 public enum ConfigurationVariable {
 
+  CANNED_ACL("com.hotels.bdp.circustrain.s3mapreducecp.cannedAcl", null),
   CREDENTIAL_PROVIDER("com.hotels.bdp.circustrain.s3mapreducecp.credentialsProvider", null),
   MINIMUM_UPLOAD_PART_SIZE("com.hotels.bdp.circustrain.s3mapreducecp.minimumUploadPartSize",
       String.valueOf(DEFAULT_TRANSFER_MANAGER_CONFIGURATION.getMinimumUploadPartSize())),