+++ circus-train-s3-s3-copier/src/main/java/com/hotels/bdp/circustrain/s3s3copier/S3S3Copier.java
@@ -202,6 +202,11 @@ public class S3S3Copier implements Copier {
 
       CopyObjectRequest copyObjectRequest = new CopyObjectRequest(s3ObjectSummary.getBucketName(),
           s3ObjectSummary.getKey(), targetS3Uri.getBucket(), targetKey);
+
+      if (s3s3CopierOptions.getCannedAcl() != null) {
+        copyObjectRequest.withCannedAccessControlList(s3s3CopierOptions.getCannedAcl());
+      }
+
       applyObjectMetadata(copyObjectRequest);
 
       TransferStateChangeListener stateChangeListener = new BytesTransferStateChangeListener(s3ObjectSummary,