+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Upgraded `hotels-oss-parent` to 2.3.5 (was 2.3.3).
 ### Added
 * Added replication-strategy configuration that can be used to support propagating deletes (drop table/partition operations). See [README.md](https://github.com/HotelsDotCom/circus-train#replication-strategy) for more details.
+* Ability to specify an S3 canned ACL via `copier-options.canned-acl`. See [#99](https://github.com/HotelsDotCom/circus-train/issues/99).
 
 ## 13.1.0 - 2018-12-20
 ### Changed